requirements.txt
dnspython
flask

